feat(frontend): changes UI
